+++ packages/@react-native-windows/automation/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @react-native-windows/automation
 
-This log was last generated on Wed, 21 Aug 2024 05:14:41 GMT and should not be manually modified.
+This log was last generated on Thu, 22 Aug 2024 05:24:28 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.3.280
+
+Thu, 22 Aug 2024 05:24:28 GMT
+
+### Patches
+
+- Bump @react-native-windows/automation-channel to v0.12.198
 
 ## 0.3.279